+++ simulation/include/Model.h
@@ -11,10 +11,10 @@
 
 #include "Geometry.h"
 
-using Meshdata = CGAL::Surface_mesh<CGAL::Simple_cartesian<double>::Point_3>;
+using Meshdata = CGAL::Surface_mesh<CGAL::Simple_cartesian<float>::Point_3>;
 
 // Will need to work on removing the headers and typedefs from this header
-typedef CGAL::Simple_cartesian<double> CoordsSystem;
+typedef CGAL::Simple_cartesian<float> CoordsSystem;
 typedef CGAL::Surface_mesh<CoordsSystem::Point_3> Mesh;
 typedef CGAL::AABB_face_graph_triangle_primitive<Mesh> Primitive;
 typedef CGAL::AABB_traits<CoordsSystem, Primitive> AABB_traits;
@@ -27,13 +27,13 @@ public:
 
     bool isModelClosed();
 
-    bool isPointInsideModel(CGAL::Simple_cartesian<double>::Point_3 point);
+    bool isPointInsideModel(CGAL::Simple_cartesian<float>::Point_3 point);
 
     const Meshdata& getMesh() const { return modelMesh; };
 
     bool checkError() { return !meshValid; };
 
-    CGAL::Simple_cartesian<double>::Iso_cuboid_3 bounding_box();
+    CGAL::Simple_cartesian<float>::Iso_cuboid_3 bounding_box();
 
     Point_3 reflectionVector(CoordsSystem::Point_3 point_1, CoordsSystem::Point_3 point_2);
