+++ simulation/source/Model.cpp
@@ -12,7 +12,7 @@
 
 #include "Model.h"
 
-typedef CGAL::Simple_cartesian<double> CoordsSystem;
+typedef CGAL::Simple_cartesian<float> CoordsSystem;
 typedef CGAL::Surface_mesh<CoordsSystem::Point_3> Mesh;
 typedef CGAL::AABB_face_graph_triangle_primitive<Mesh> Primitive;
 typedef CGAL::AABB_traits<CoordsSystem, Primitive> AABB_traits;
@@ -48,7 +48,7 @@ bool Model::isModelClosed()
     return CGAL::is_closed(modelMesh);
 }
 
-bool Model::isPointInsideModel(CGAL::Simple_cartesian<double>::Point_3 point)
+bool Model::isPointInsideModel(CGAL::Simple_cartesian<float>::Point_3 point)
 {
     if(!meshValid)
     {
@@ -67,7 +67,7 @@ void Model::lazyAabbTreeConstruction()
     modelTree = new AABB_tree(faces(modelMesh).first, faces(modelMesh).second, modelMesh);
 }
 
-CGAL::Simple_cartesian<double>::Iso_cuboid_3 Model::bounding_box()
+CGAL::Simple_cartesian<float>::Iso_cuboid_3 Model::bounding_box()
 {
     return CGAL::bounding_box(modelMesh.points().begin(), modelMesh.points().end());
 }
